renaming generalized poisson llh modules
