+++ pisa/stages/simple_fit/super_simple_pipeline.cfg
@@ -4,7 +4,7 @@ super_simple_binning.order = stuff
 super_simple_binning.stuff = {'bin_edges':np.linspace(0.,40.,51)}
 
 
-[simple_fit.super_simple_signal]
+[simple_fit.pi_simple_signal]
 
 input_specs  = None
 calc_specs   = None
@@ -25,7 +25,7 @@ param.sigma.fixed = True
 
 
 
-[test_bourdeet.prepare_generalized_llh_parameters]
+[likelihood.pi_generalized_llh_params]
 
 
 input_specs = None
@@ -34,7 +34,6 @@ output_specs = super_simple_binning
 
 
 
-
 [pipeline]
 
-order = simple_fit.super_simple_signal, test_bourdeet.prepare_generalized_llh_parameters
+order = simple_fit.pi_simple_signal, likelihood.pi_generalized_llh_params
